backend/requirements.txt
aerich==0.7.2
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
ariadne==0.23.0
ariadne-graphql-modules==0.8.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.7.4
click==8.1.7
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-sessions==0.3.2
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tomlkit==0.13.0
tortoise-orm==0.21.5
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
